Move hyperparameter optimization stuff to hparam_search.py
