+++ train.py
@@ -3,7 +3,6 @@
 from abc import ABC, abstractmethod
 from collections import Counter, deque
 from typing import NamedTuple, Literal, Any
-import os
 import torch 
 import numpy as np
 import tqdm
@@ -13,10 +12,7 @@ import pandas as pd
 
 from model import CheckersQModel, CheckersVModel
 from game import Game, GameState, Player, Action, NUM_ACTIONS
-from agent import play_agent_game, BaseAgent, QModelAgent, RandomAgent, UserInputAgent, VModelAgent
-from skopt import BayesSearchCV
-from skopt.space import Integer, Real
-from sklearn.base import BaseEstimator, RegressorMixin
+from agent import play_agent_game, BaseAgent, QModelAgent, RandomAgent, VModelAgent
 
 
 class Sample(NamedTuple):
@@ -312,128 +308,10 @@ class VModelTrainRun(TrainRun):
         return losses
 
 
-@torch.no_grad()
-def evaluate_model_vs_random(rl_agent: QModelAgent, random_agent: RandomAgent, epsilon: float, num_games: int):
-    game = Game()
-    player_wins = 0
-    draws = 0
-
-    for _ in tqdm.trange(num_games, position=1, leave=False, desc='Playing against random moves'):
-        rl_player, winner, _ = play_agent_game(game, rl_agent, random_agent, dict(epsilon=epsilon))
-        if winner == rl_player:
-            player_wins += 1
-        elif winner == Player.NEUTRAL:
-            draws += 1
-
-    win_rate = player_wins / num_games
-    draw_rate = draws / num_games
-
-    return win_rate, draw_rate
-
-
-class CheckersQModelWrapper(BaseEstimator, RegressorMixin):
-    def __init__(self, num_hidden_layers=0, hidden_size=0, learning_rate=0.01,
-                 self_play_games_per_iter=25, weight_decay=1e-4):
-        self.num_hidden_layers = num_hidden_layers
-        self.hidden_size = hidden_size
-        self.learning_rate = learning_rate
-        self.self_play_games_per_iter = self_play_games_per_iter
-        self.weight_decay = weight_decay
-
-        # Initialize model here
-        self.model = None
-
-        # Initialize train_hist to an empty list
-        self.train_hist = []
-
-    def fit(self, X, y=None, **kwargs):
-        # Update the model's parameters
-        self.model = CheckersQModel(
-            num_hidden_layers=self.num_hidden_layers,
-            hidden_size=self.hidden_size
-        )
-
-        # Set other parameters as needed
-        self.optimzer = torch.optim.SGD(self.model.parameters(),
-                                        lr=self.learning_rate,
-                                        weight_decay=self.weight_decay)
-        self.train_run = QModelTrainRun(self.model, self.optimzer)
-
-        # Train the model and get the metric to optimize (e.g., average win rate)
-        self.train_hist = self.train_run.train(selfplay_games_p_i=self.self_play_games_per_iter)
-        return self
-
-    def score(self, X, y=None, **kwargs):
-        num_evaluation_games = 100
-        wins, draws = self.train_run.evaluate_strength(num_evaluation_games=num_evaluation_games)
-        return wins / num_evaluation_games
-
-    def get_params(self, deep=True):
-        # Implement get_params method here
-        return {'num_hidden_layers': self.num_hidden_layers,
-                'hidden_size': self.hidden_size,
-                'learning_rate': self.learning_rate,
-                'self_play_games_per_iter': self.self_play_games_per_iter,
-                'weight_decay': self.weight_decay}
-
-    def set_params(self, **parameters):
-        # Implement set_params method here
-        for parameter, value in parameters.items():
-            setattr(self, parameter, value)
-        return self
-
-
-def optimize_hyperparameters():
-    # Define the hyperparameter search space
-    search_space = {
-        'num_hidden_layers': Integer(0, 3),
-        'hidden_size': Integer(64, 1024),
-        'learning_rate': Real(1e-5, 1e-1, 'log-uniform'),
-        'self_play_games_per_iter': Integer(1, 25),
-        'weight_decay': Real(1e-5, 1e-1, 'log-uniform')
-    }
-
-    wrapper = CheckersQModelWrapper()
-
-    # Use BayesSearchCV for optimization
-    opt = BayesSearchCV(wrapper, search_space, n_iter=30, random_state=42, verbose=2)
-
-    # Pass a dummy X (input data) and y (target) for optimization
-    X_dummy = np.random.rand(100, 10)  # 100 samples, 10 features
-    y_dummy = np.random.randint(0, 2, 100)
-    best_optimized_params = opt.fit(X_dummy, y_dummy).best_params_
-
-    return best_optimized_params
-
-
-
-
-
 if __name__ == '__main__':
-    # model = CheckersQModel(num_hidden_layers=1, hidden_size=256)
-    # optimizer = torch.optim.SGD(model.parameters(), lr=1e-3, weight_decay=3.0)
-    # trainrun = QModelTrainRun(model, optimizer)
-    # train_hist = trainrun.train()
-    # print(train_hist)
+    model = CheckersQModel(num_hidden_layers=1, hidden_size=256)
+    optimizer = torch.optim.SGD(model.parameters(), lr=1e-3, weight_decay=3.0)
+    trainrun = QModelTrainRun(model, optimizer)
+    train_hist = trainrun.train()
+    print(train_hist)
 
-    # Perform Bayesian optimization to find the best hyperparameters
-    best_hyperparams = optimize_hyperparameters()
-
-    # Use the best hyperparameters to train the model
-    best_model = CheckersQModel(
-        num_hidden_layers=best_hyperparams['num_hidden_layers'],
-        hidden_size=best_hyperparams['hidden_size']
-    )
-    import pdb; pdb.set_trace()
-
-    # Train the model using the best hyperparameters
-    # train_hist = train_loop(
-        # best_model,
-        # learning_rate=best_hyperparams['learning_rate'],
-        # self_play_games_per_iter=best_hyperparams['self_play_games_per_iter'],
-        # weight_decay=best_hyperparams['weight_decay']
-    # )
-
-    # Evaluate the model against random moves with the best hyperparameters
-    # win_rate, draw_rate = evaluate_model_vs_random(best_model, epsilon=0.05, num_games=10)
-    # print(f'Final Win Rate: {win_rate}, Draw Rate: {draw_rate}')
